Allow dynamic icon mappings
This feature builds upon the idea in #15: for some kinds of icons, it'd
be best to know the candidate itself, not just the kind. The simplest
example is files, like in that PR, which makes the icon match the
filename and extension of the path. Instead of single-casing, this
commit adds a standard way to do this, and makes it available to the
user. Currently, it's used for directories along with files.