Refresh space constant on frame creation
This fixes #16. It happened because the `defconst` form was being
evaluated at startup, but when Emacs starts as a headless server, that's
a value that is only valid for terminal frames. Opening a graphical
frame didn't re-evaluate the value, so the popup would show clipping.
Now we reset the constant on the creation of any server-connecting
frames.